+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Since 0.19.0-beta, we use [this changelog format](https://keepachangelog.com). I
 * Key `"run_noug_dir_work_dir"` in BuiltinFunction dictionnaries (and in libs too) was renamed to `"run_noug_dir"`.
 * Type of functions is now `func`, regardless of if they are functions or built-in functions. However, type of methods is still `method`.
 * Modules and functions are now True in boolean context (i.e. in a condition)
+* `round` built-in function now returns `int` if `n_digits` ⩽ 0.
 * `(Internal API)` `dived_by` and `floor_dived_by` methods in Values are now renamed to `divided_by` and `floor_divided_by`.
 * `(Build scripts)` (Windows) The Nougaro version is now automatically fetched.